SVM using PCA
This code takes a set of labelled images loaded from a local source, and performs PCA on them.  This simplified input is then fed to an SVC which attempts to classify test images into one of three categories.  The results are somewhat underwhelming, achieving on average a 53% accuracy rate.  While the input data are quite lousy (e.g. many pictures of dogs have the dogs as secondary features) and limited in number, the CNN code also on my github does a far better job (76%).  I suspect that throwing away the relative pixel position information as this code does is too costly an approximation for this problem.